requirements.txt
# Filename: requirements.txt
weasyprint==62.2
cssutils==2.11.1
mistune==3.0.2
PyYAML==6.0.2
PyQt6==6.7.0
PyQt6-WebEngine==6.7.0
pytest==8.3.3